+++ arch/arm32/init.c
@@ -28,6 +28,8 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #define UNUSED_ARG __attribute__((unused))
 
 void arch_kmain(UNUSED_ARG uint32_t r0, UNUSED_ARG uint32_t r1, UNUSED_ARG uint32_t dtb_ptr) {
+    early_kmain();
+
     init_pmm();
     init_vmm();